Added fb cover_image to parse user
